Add tests for estimateGas with precompiles related to associate, dissociate, approve, transfer

+++ hedera-mirror-test/src/test/java/com/hedera/mirror/test/e2e/acceptance/client/TokenClient.java
@@ -90,7 +90,7 @@ public class TokenClient extends AbstractNetworkClient {
             int freezeStatus,
             int kycStatus,
             ExpandedAccountId treasuryAccount,
-            int initialSupply,
+            long initialSupply,
             TokenSupplyType tokenSupplyType,
             long maxSupply,
             TokenType tokenType,
@@ -177,20 +177,20 @@ public class TokenClient extends AbstractNetworkClient {
             int freezeStatus,
             int kycStatus,
             ExpandedAccountId treasuryAccount,
-            int initialSupply,
+            long initialSupply,
             TokenSupplyType tokenSupplyType,
             long maxSupply,
             List<CustomFee> customFees) {
         TokenCreateTransaction tokenCreateTransaction = getTokenCreateTransaction(
-                        expandedAccountId,
-                        symbol,
-                        freezeStatus,
-                        kycStatus,
-                        treasuryAccount,
-                        TokenType.FUNGIBLE_COMMON,
-                        tokenSupplyType,
-                        maxSupply,
-                        customFees)
+                expandedAccountId,
+                symbol,
+                freezeStatus,
+                kycStatus,
+                treasuryAccount,
+                TokenType.FUNGIBLE_COMMON,
+                tokenSupplyType,
+                maxSupply,
+                customFees)
                 .setDecimals(10)
                 .setInitialSupply(initialSupply);
